Handle automatic PWA updates when new deployment is made
