+++ apps/client/src/enums/cache.ts
@@ -1,4 +1,3 @@
 export enum ServiceWorkerCache {
   APPLICATION = 'application',
-  RUNTIME = 'runtime',
 }
