+++ apps/client/src/enums/event.ts
@@ -13,5 +13,5 @@ export enum ServiceWorkerEvent {
   APPLICATION_CACHE_INIT = 'APPLICATION_CACHE_INIT',
   APPLICATION_CACHE_FINISHED = 'APPLICATION_CACHE_FINISHED',
 
-  RUNTIME_CACHE_REQUEST_FAILED = 'RUNTIME_CACHE_REQUEST_FAILED',
+  APPLICATION_CACHE_REQUEST_FAILED = 'APPLICATION_CACHE_REQUEST_FAILED',
 }
